requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.1
sentence-transformers==2.2.2
chromadb==0.4.15
openai==1.3.0
python-dotenv==1.0.0
gradio==3.50.0
streamlit==1.28.0
huggingface-hub==0.19.0